refactor: css 분리, 리팩토링 (공통, 컴포넌트, 변수화)
