--- css/common.css
@@ -1,176 +1,0 @@
-* {
-  box-sizing: border-box !important;
-}
-
-body {
-  font-size: 14px;
-  background-color: #222222;
-  color: #fff;
-}
-
-a {
-  color: inherit;
-  text-decoration: none;
-}
-
-button {
-  cursor: pointer;
-}
-
-#app {
-  padding-bottom: 48px;
-}
-
-*:focus {
-  outline: none;
-}
-.item-view,
-.item-test {
-  width: 100%;
-}
-
-.item-view {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 1200px;
-  margin: 0 auto;
-}
-
-.item-view h2 {
-  font-size: 2rem;
-  font-weight: bold;
-  user-select: none;
-}
-
-/* https://andrew.hedges.name/experiments/aspect_ratio/ */
-
-.item-list {
-  display: grid;
-  margin: 48px 0;
-  grid-template-columns: repeat(4, 180px);
-  grid-column-gap: 160px;
-  grid-row-gap: 48px;
-}
-
-.item-card {
-  display: flex;
-  flex-direction: column;
-}
-
-.item-thumbnail {
-  border-radius: 8px;
-  width: 180px;
-  height: 270px;
-  background-size: contain;
-}
-
-.item-title {
-  margin-top: 16px;
-  font-size: 1.2rem;
-  font-weight: bold;
-}
-
-.item-score {
-  margin-top: 16px;
-  font-size: 1.2rem;
-  display: inline-block;
-  vertical-align: middle;
-}
-
-.star-start {
-  margin-left: 5px;
-  vertical-align: middle;
-}
-
-.item-score::after {
-  margin-left: 8px;
-}
-
-.item-title.skeleton::after,
-.item-score.skeleton::after {
-  font-size: 0;
-  content: 'loading';
-}
-
-.full-width {
-  width: 100%;
-}
-
-.last-item {
-  margin-top: 48px;
-}
-
-button.btn {
-  border: 0;
-  border-radius: 8px;
-  height: 30px;
-  color: #fff;
-}
-
-button.primary {
-  background: #f33f3f;
-}
-
-.item-card .skeleton {
-  background: linear-gradient(-90deg, #aaa, #f0f0f0, #aaa, #f0f0f0);
-  background-size: 400%;
-  animation: skeleton-animation 5s infinite ease-out;
-  border-radius: 8px;
-}
-
-@keyframes skeleton-animation {
-  0% {
-    background-position: 0% 50%;
-  }
-  50% {
-    background-position: 100% 50%;
-  }
-  100% {
-    background-position: 0% 50%;
-  }
-}
-header {
-  width: 100%;
-  min-width: 1200px;
-  height: 72px;
-  background-color: #222;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 20px;
-  border-bottom: 1px solid white;
-  margin-bottom: 48px;
-}
-
-header h1 {
-  cursor: pointer;
-  user-select: none;
-  font-size: 2rem;
-  font-weight: bold;
-  letter-spacing: -0.1rem;
-  color: #f33f3f;
-}
-
-#home-button {
-  background-color: transparent;
-  border: none;
-}
-
-header > .search-box {
-  background: #fff;
-  padding: 8px;
-  border-radius: 4px;
-}
-
-header .search-box > input {
-  border: 0;
-}
-
-header .search-box > .search-button {
-  width: 14px;
-  border: 0;
-  text-indent: -1000rem;
-  background: url('../src/images/search_button.png') transparent no-repeat 0 1px;
-  background-size: contain;
-}
