+++ docs/REFACTOR.md
@@ -1,0 +1,9 @@
+## 1단계 피드백 후 반영
+
+- [ ] E2E 네트워크 상태나 응답을 시뮬레이션
+- [ ] E2E 에러에 대한 테스트
+
+- [x] css 분리, 리팩토링 (공통, 컴포넌트, 변수화)
+- [ ] 코드의 일관성
+
+- [ ] 무한스크롤 구현하며 API에서 몇개의 데이터가 올지 예상, Skeleton 구현
