+++ src/index.js
@@ -1,6 +1,5 @@
 import App from './App';
-import '../css/reset.css';
-import '../css/common.css';
+import './index.css';
 
 const app = new App();
